Rename `regular -> quote`
