+++ crates/shared/src/price_estimation/competition/quote.rs
